split collection
