+++ packages/mongo/collection/methods_sync.js
@@ -1,0 +1,346 @@
+export const SyncMethods = {
+  /**
+   * @summary Find the documents in a collection that match the selector.
+   * @locus Anywhere
+   * @method find
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {MongoSelector} [selector] A query describing the documents to find
+   * @param {Object} [options]
+   * @param {MongoSortSpecifier} options.sort Sort order (default: natural order)
+   * @param {Number} options.skip Number of results to skip at the beginning
+   * @param {Number} options.limit Maximum number of results to return
+   * @param {MongoFieldSpecifier} options.fields Dictionary of fields to return or exclude.
+   * @param {Boolean} options.reactive (Client only) Default `true`; pass `false` to disable reactivity
+   * @param {Function} options.transform Overrides `transform` on the  [`Collection`](#collections) for this cursor.  Pass `null` to disable transformation.
+   * @param {Boolean} options.disableOplog (Server only) Pass true to disable oplog-tailing on this query. This affects the way server processes calls to `observe` on this query. Disabling the oplog can be useful when working with data that updates in large batches.
+   * @param {Number} options.pollingIntervalMs (Server only) When oplog is disabled (through the use of `disableOplog` or when otherwise not available), the frequency (in milliseconds) of how often to poll this query when observing on the server. Defaults to 10000ms (10 seconds).
+   * @param {Number} options.pollingThrottleMs (Server only) When oplog is disabled (through the use of `disableOplog` or when otherwise not available), the minimum time (in milliseconds) to allow between re-polling when observing on the server. Increasing this will save CPU and mongo load at the expense of slower updates to users. Decreasing this is not recommended. Defaults to 50ms.
+   * @param {Number} options.maxTimeMs (Server only) If set, instructs MongoDB to set a time limit for this cursor's operations. If the operation reaches the specified time limit (in milliseconds) without the having been completed, an exception will be thrown. Useful to prevent an (accidental or malicious) unoptimized query from causing a full collection scan that would disrupt other database users, at the expense of needing to handle the resulting error.
+   * @param {String|Object} options.hint (Server only) Overrides MongoDB's default index selection and query optimization process. Specify an index to force its use, either by its name or index specification. You can also specify `{ $natural : 1 }` to force a forwards collection scan, or `{ $natural : -1 }` for a reverse collection scan. Setting this is only recommended for advanced users.
+   * @param {String} options.readPreference (Server only) Specifies a custom MongoDB [`readPreference`](https://docs.mongodb.com/manual/core/read-preference) for this particular cursor. Possible values are `primary`, `primaryPreferred`, `secondary`, `secondaryPreferred` and `nearest`.
+   * @returns {Mongo.Cursor}
+   */
+  find(...args) {
+    // Collection.find() (return all docs) behaves differently
+    // from Collection.find(undefined) (return 0 docs).  so be
+    // careful about the length of arguments.
+    return this._collection.find(
+      this._getFindSelector(args),
+      this._getFindOptions(args)
+    );
+  },
+
+  /**
+   * @summary Finds the first document that matches the selector, as ordered by sort and skip options. Returns `undefined` if no matching document is found.
+   * @locus Anywhere
+   * @method findOne
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {MongoSelector} [selector] A query describing the documents to find
+   * @param {Object} [options]
+   * @param {MongoSortSpecifier} options.sort Sort order (default: natural order)
+   * @param {Number} options.skip Number of results to skip at the beginning
+   * @param {MongoFieldSpecifier} options.fields Dictionary of fields to return or exclude.
+   * @param {Boolean} options.reactive (Client only) Default true; pass false to disable reactivity
+   * @param {Function} options.transform Overrides `transform` on the [`Collection`](#collections) for this cursor.  Pass `null` to disable transformation.
+   * @param {String} options.readPreference (Server only) Specifies a custom MongoDB [`readPreference`](https://docs.mongodb.com/manual/core/read-preference) for fetching the document. Possible values are `primary`, `primaryPreferred`, `secondary`, `secondaryPreferred` and `nearest`.
+   * @returns {Object}
+   */
+  findOne(...args) {
+    return this._collection.findOne(
+      this._getFindSelector(args),
+      this._getFindOptions(args)
+    );
+  },
+
+
+  // 'insert' immediately returns the inserted document's new _id.
+  // The others return values immediately if you are in a stub, an in-memory
+  // unmanaged collection, or a mongo-backed collection and you don't pass a
+  // callback. 'update' and 'remove' return the number of affected
+  // documents. 'upsert' returns an object with keys 'numberAffected' and, if an
+  // insert happened, 'insertedId'.
+  //
+  // Otherwise, the semantics are exactly like other methods: they take
+  // a callback as an optional last argument; if no callback is
+  // provided, they block until the operation is complete, and throw an
+  // exception if it fails; if a callback is provided, then they don't
+  // necessarily block, and they call the callback when they finish with error and
+  // result arguments.  (The insert method provides the document ID as its result;
+  // update and remove provide the number of affected docs as the result; upsert
+  // provides an object with numberAffected and maybe insertedId.)
+  //
+  // On the client, blocking is impossible, so if a callback
+  // isn't provided, they just return immediately and any error
+  // information is lost.
+  //
+  // There's one more tweak. On the client, if you don't provide a
+  // callback, then if there is an error, a message will be logged with
+  // Meteor._debug.
+  //
+  // The intent (though this is actually determined by the underlying
+  // drivers) is that the operations should be done synchronously, not
+  // generating their result until the database has acknowledged
+  // them. In the future maybe we should provide a flag to turn this
+  // off.
+
+  _insert(doc, callback) {
+    // Make sure we were passed a document to insert
+    if (!doc) {
+      throw new Error('insert requires an argument');
+    }
+
+
+    // Make a shallow clone of the document, preserving its prototype.
+    doc = Object.create(
+      Object.getPrototypeOf(doc),
+      Object.getOwnPropertyDescriptors(doc)
+    );
+
+    if ('_id' in doc) {
+      if (
+        !doc._id ||
+        !(typeof doc._id === 'string' || doc._id instanceof Mongo.ObjectID)
+      ) {
+        throw new Error(
+          'Meteor requires document _id fields to be non-empty strings or ObjectIDs'
+        );
+      }
+    } else {
+      let generateId = true;
+
+      // Don't generate the id if we're the client and the 'outermost' call
+      // This optimization saves us passing both the randomSeed and the id
+      // Passing both is redundant.
+      if (this._isRemoteCollection()) {
+        const enclosing = DDP._CurrentMethodInvocation.get();
+        if (!enclosing) {
+          generateId = false;
+        }
+      }
+
+      if (generateId) {
+        doc._id = this._makeNewID();
+      }
+    }
+
+
+    // On inserts, always return the id that we generated; on all other
+    // operations, just return the result from the collection.
+    var chooseReturnValueFromCollectionResult = function(result) {
+      if (Meteor._isPromise(result)) return result;
+
+      if (doc._id) {
+        return doc._id;
+      }
+
+      // XXX what is this for??
+      // It's some iteraction between the callback to _callMutatorMethod and
+      // the return value conversion
+      doc._id = result;
+
+      return result;
+    };
+
+    const wrappedCallback = wrapCallback(
+      callback,
+      chooseReturnValueFromCollectionResult
+    );
+
+    if (this._isRemoteCollection()) {
+      const result = this._callMutatorMethod('insert', [doc], wrappedCallback);
+      return chooseReturnValueFromCollectionResult(result);
+    }
+
+    // it's my collection.  descend into the collection object
+    // and propagate any exception.
+    try {
+      // If the user provided a callback and the collection implements this
+      // operation asynchronously, then queryRet will be undefined, and the
+      // result will be returned through the callback instead.
+      let result;
+      if (!!wrappedCallback) {
+        this._collection.insert(doc, wrappedCallback);
+      } else {
+        // If we don't have the callback, we assume the user is using the promise.
+        // We can't just pass this._collection.insert to the promisify because it would lose the context.
+        result = this._collection.insert(doc);
+      }
+
+      return chooseReturnValueFromCollectionResult(result);
+    } catch (e) {
+      if (callback) {
+        callback(e);
+        return null;
+      }
+      throw e;
+    }
+  },
+
+  /**
+   * @summary Insert a document in the collection.  Returns its unique _id.
+   * @locus Anywhere
+   * @method  insert
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {Object} doc The document to insert. May not yet have an _id attribute, in which case Meteor will generate one for you.
+   * @param {Function} [callback] Optional.  If present, called with an error object as the first argument and, if no error, the _id as the second.
+   */
+  insert(doc, callback) {
+    return this._insert(doc, callback);
+  },
+
+  /**
+   * @summary Asynchronously modifies one or more documents in the collection. Returns the number of matched documents.
+   * @locus Anywhere
+   * @method update
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {MongoSelector} selector Specifies which documents to modify
+   * @param {MongoModifier} modifier Specifies how to modify the documents
+   * @param {Object} [options]
+   * @param {Boolean} options.multi True to modify all matching documents; false to only modify one of the matching documents (the default).
+   * @param {Boolean} options.upsert True to insert a document if no matching documents are found.
+   * @param {Array} options.arrayFilters Optional. Used in combination with MongoDB [filtered positional operator](https://docs.mongodb.com/manual/reference/operator/update/positional-filtered/) to specify which elements to modify in an array field.
+   * @param {Function} [callback] Optional.  If present, called with an error object as the first argument and, if no error, the number of affected documents as the second.
+   */
+  update(selector, modifier, ...optionsAndCallback) {
+    const callback = popCallbackFromArgs(optionsAndCallback);
+
+    // We've already popped off the callback, so we are left with an array
+    // of one or zero items
+    const options = { ...(optionsAndCallback[0] || null) };
+    let insertedId;
+    if (options && options.upsert) {
+      // set `insertedId` if absent.  `insertedId` is a Meteor extension.
+      if (options.insertedId) {
+        if (
+          !(
+            typeof options.insertedId === 'string' ||
+            options.insertedId instanceof Mongo.ObjectID
+          )
+        )
+          throw new Error('insertedId must be string or ObjectID');
+        insertedId = options.insertedId;
+      } else if (!selector || !selector._id) {
+        insertedId = this._makeNewID();
+        options.generatedId = true;
+        options.insertedId = insertedId;
+      }
+    }
+
+    selector = Mongo.Collection._rewriteSelector(selector, {
+      fallbackId: insertedId,
+    });
+
+    const wrappedCallback = wrapCallback(callback);
+
+    if (this._isRemoteCollection()) {
+      const args = [selector, modifier, options];
+      return this._callMutatorMethod('update', args, callback);
+    }
+
+    // it's my collection.  descend into the collection object
+    // and propagate any exception.
+    // If the user provided a callback and the collection implements this
+    // operation asynchronously, then queryRet will be undefined, and the
+    // result will be returned through the callback instead.
+    //console.log({callback, options, selector, modifier, coll: this._collection});
+    try {
+      // If the user provided a callback and the collection implements this
+      // operation asynchronously, then queryRet will be undefined, and the
+      // result will be returned through the callback instead.
+      return this._collection.update(
+        selector,
+        modifier,
+        options,
+        wrappedCallback
+      );
+    } catch (e) {
+      if (callback) {
+        callback(e);
+        return null;
+      }
+      throw e;
+    }
+  },
+
+  /**
+   * @summary Remove documents from the collection
+   * @locus Anywhere
+   * @method remove
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {MongoSelector} selector Specifies which documents to remove
+   * @param {Function} [callback] Optional.  If present, called with an error object as the first argument and, if no error, the number of affected documents as the second.
+   */
+  remove(selector, callback) {
+    selector = Mongo.Collection._rewriteSelector(selector);
+
+    if (this._isRemoteCollection()) {
+      return this._callMutatorMethod('remove', [selector], callback);
+    }
+
+
+    // it's my collection.  descend into the collection1 object
+    // and propagate any exception.
+    return this._collection.remove(selector);
+  },
+
+  /**
+   * @summary Asynchronously modifies one or more documents in the collection, or insert one if no matching documents were found. Returns an object with keys `numberAffected` (the number of documents modified)  and `insertedId` (the unique _id of the document that was inserted, if any).
+   * @locus Anywhere
+   * @method upsert
+   * @memberof Mongo.Collection
+   * @instance
+   * @param {MongoSelector} selector Specifies which documents to modify
+   * @param {MongoModifier} modifier Specifies how to modify the documents
+   * @param {Object} [options]
+   * @param {Boolean} options.multi True to modify all matching documents; false to only modify one of the matching documents (the default).
+   * @param {Function} [callback] Optional.  If present, called with an error object as the first argument and, if no error, the number of affected documents as the second.
+   */
+  upsert(selector, modifier, options, callback) {
+    if (!callback && typeof options === 'function') {
+      callback = options;
+      options = {};
+    }
+
+    return this.update(
+      selector,
+      modifier,
+      {
+        ...options,
+        _returnObject: true,
+        upsert: true,
+      });
+  },
+}
+
+// Convert the callback to not return a result if there is an error
+function wrapCallback(callback, convertResult) {
+  return (
+    callback &&
+    function(error, result) {
+      if (error) {
+        callback(error);
+      } else if (typeof convertResult === 'function') {
+        callback(error, convertResult(result));
+      } else {
+        callback(error, result);
+      }
+    }
+  );
+}
+
+function popCallbackFromArgs(args) {
+  // Pull off any callback (or perhaps a 'callback' variable that was passed
+  // in undefined, like how 'upsert' does it).
+  if (
+    args.length &&
+    (args[args.length - 1] === undefined ||
+      args[args.length - 1] instanceof Function)
+  ) {
+    return args.pop();
+  }
+}
