+++ packages/mongo/package.js
@@ -101,8 +101,8 @@ Package.onUse(function (api) {
   );
   api.addFiles("local_collection_driver.js", ["client", "server"]);
   api.addFiles("remote_collection_driver.ts", "server");
-  api.addFiles("collection.js", ["client", "server"]);
-  api.addFiles("connection_options.js", "server");
+  api.addFiles("collection/collection.js", ["client", "server"]);
+  api.addFiles("connection_options.ts", "server");
   api.addAssets("mongo.d.ts", "server");
 });
